fix: API경로, 타입, 생성자 주입
