+++ src/main/java/knu/kproject/dto/UserDto/TeamUserInfo.java
@@ -6,18 +6,12 @@ import lombok.Data;
 import java.util.List;
 
 @Data
-public class AdminUserDto {
-    private Long id;
+public class TeamUserInfo {
     private String name;
     private String socialId;
-    private boolean requiredTermsAgree;
-    private boolean marketingEmailOptIn;
     private String phone;
     private String address;
     private String mbti;
-    private String status;
-    private String oauth2Id;
-    private String role;
     private List<ToolDto> tools;
-    private String profileImage;
+    private String imagePath;
 }
