+++ src/main/java/knu/kproject/dto/UserDto/ToolDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ToolDto {
-    private Integer id;
+    private Long id;
     private String name;
     private String nameId;
 }
