+++ src/main/java/knu/kproject/dto/UserDto/UserDto.java
@@ -14,6 +14,6 @@ public class UserDto {
     private String address;
     private String mbti;
     private List<ToolDto> tools;
-    private String profileImage;
+    private String imagePath;
 
 }
